SMS/SMS/requirements.txt
pillow
django-crispy-forms
foundation
six
django-localflavor
django-phonenumber-field
 django-countries
 twilio==6.35.4
